+++ platform/core/src/stats/finders.ts
@@ -22,7 +22,6 @@ import {
   followsMapping,
   postsWithActivityScoreMapping,
   profileWithActivityScoreMapping,
-  reactionsMapping,
   repostsMapping,
 } from './mapping';
 import { compileStatsWithAll, createdAtFilter } from './helpers';
@@ -32,7 +31,7 @@ import {
 } from '../jams/mapping';
 import { edgeFilters, postFilters, profileFilters } from '../db/pg/filters';
 import { sorts } from '../db/pg/queries';
-import { entries, follows, reactions, repost } from '../db/pg/schema/edges';
+import { entries, follows, repost } from '../db/pg/schema/edges';
 import { jamEvents, posts, profiles } from '../db/pg/schema/documents';
 
 export const activeProfilesCount = (db: PgDb, start?: Date, end?: Date) =>
@@ -83,7 +82,10 @@ export const postsRepliesStats = (): Promise<ObjectStatsWithAll> =>
   compileStatsWithAll(postsRepliesCount);
 
 const reactionCount = (db: PgDb, start?: Date, end?: Date) =>
-  reactionsMapping.filter(createdAtFilter(reactions, start, end)).count(db);
+  entriesMapping
+    .filter(createdAtFilter(entries, start, end))
+    .filter(edgeFilters.entryType('reaction'))
+    .count(db);
 const rsvpCount = (db: PgDb, start?: Date, end?: Date) =>
   entriesMapping
     .filter(createdAtFilter(entries, start, end))
@@ -97,7 +99,6 @@ const repostCount = (db: PgDb, start?: Date, end?: Date) =>
   repostsMapping.filter(createdAtFilter(repost, start, end)).count(db);
 
 const interactionsCount = async (db: PgDb, start?: Date, end?: Date) =>
-  (await reactionCount(db, start, end)) +
   (await followCount(db, start, end)) +
   (await entryCount(db, start, end)) +
   (await repostCount(db, start, end));